GBLD-694: enrich toolchains.xml with Jenkins's JDK installations
The idea is to provide one or more Jenkins-wide file templates for
toolchains.xml.

The template contains preconfigured jdk toolchain entries with an id
that is equal to Jenkins's JDK tool configuration.

In the toolchains.xml template, these entries are adjusted to use the
javaHome from Jenkins's JDK tool.

If there are multiple JDK installations for the same java version in the
template, Maven can reference the desired one by id.

If Jenkins's JDK tools cannot be found in the toolchains.xml template,
they are not added to the toolchains.xml and a warning is printed per
JDK tool.