DINF-161: fix versioning
This version is currently based on upstream 3.11.1. However, 3.11.1 hardcodes version to 3.11.1, instead of dynamic ${revision}${changelist}, breaking our snapshot jobs (which can't build release versions).

So we set version back to ${revision}${changelist} and set revision to 3.11.1 so that our next release will become 3.11.1-gebit1.

When we merge in release versions, we get